requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
asgiref==3.8.1
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.3.2
cryptography==43.0.1
decorator==4.4.2
defusedxml==0.8.0rc2
Deprecated==1.2.14
distro==1.9.0
Django==5.1.1
django-templated-mail==1.1.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
djoser==2.2.3
et_xmlfile==2.0.0
exceptiongroup==1.2.2
fonttools==4.54.1
fpdf==1.7.2
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
imageio==2.36.0
imageio-ffmpeg==0.5.1
jiter==0.7.0
lxml==5.3.0
moviePy
numpy==2.1.2
oauthlib==3.2.2
openai==1.53.0
opencv-python-headless==4.10.0.84
openpyxl==3.1.5
packaging==24.1
pandas==2.2.3
pdf2docx==0.5.8
pikepdf==9.4.0
pillow==10.4.0
pillow_heif==0.20.0
proglog==0.1.10
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.9.0
PyMuPDF==1.24.13
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-docx==1.1.2
python3-openid==3.2.0
pytz==2024.2
rawpy==0.23.2
reportlab==4.2.5
requests==2.32.3
requests-oauthlib==2.0.0
six==1.16.0
sniffio==1.3.1
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.1
termcolor==2.5.0
tqdm==4.66.6
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
wrapt==1.16.0
xlrd==2.0.1
